fix iOS large numbers

+++ lib/shaders/lineFrag.glsl
@@ -1,4 +1,5 @@
-precision mediump float;
+precision highp float;
+
 uniform vec4 color;
 void main() {
   gl_FragColor = color;
